add register controller, users controller etc.

+++ src/main/java/my/notes/notesApp/biz/service/CustomerService.java
@@ -2,6 +2,7 @@ package my.notes.notesApp.biz.service;
 
 import lombok.extern.log4j.Log4j2;
 import my.notes.notesApp.biz.model.Customer;
+import my.notes.notesApp.biz.model.Note;
 import my.notes.notesApp.biz.model.Role;
 import my.notes.notesApp.data.CustomerRepository;
 import my.notes.notesApp.data.NoteRepository;
@@ -14,9 +15,11 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.nio.file.attribute.UserPrincipalNotFoundException;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Log4j2
 @Service
@@ -39,7 +42,7 @@ public class CustomerService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        List<Customer> customers = customerRepository.findByUserName(username);
+        List<Customer> customers = customerRepository.findByUsername(username);
         if (customers.isEmpty()) {
             throw new UsernameNotFoundException("User not found with username: " + username);
         }
@@ -47,9 +50,9 @@ public class CustomerService implements UserDetailsService {
     }
 
     public UserDetails createNewCustomer(String username, String email, String password) {
-        if (customerRepository.findByUserName(username).isEmpty()) {
+        if (customerRepository.findByUsername(username).isEmpty()) {
             Customer customer = Customer.builder()
-                    .userName(username)
+                    .username(username)
                     .email(email)
                     .password(passwordEncoder.encode(password))
                     .roles(new HashSet<>())
@@ -64,7 +67,7 @@ public class CustomerService implements UserDetailsService {
     }
 
     public void addRoleToCustomer(String username, String roleName) {
-        Customer customer = customerRepository.findByUserName(username).getFirst();
+        Customer customer = customerRepository.findByUsername(username).getFirst();
         Optional<Role> role = roleRepository.findByName(roleName);
         if (roleRepository.findByName(roleName).isPresent() && customer != null) {
             customer.getRoles().add(role.get());
@@ -82,17 +85,62 @@ public class CustomerService implements UserDetailsService {
 
     public Customer getCurrentUser() {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
-        return customerRepository.findByUserName(username).getFirst(); // TODO: think about handling exception here
+        return customerRepository.findByUsername(username).getFirst(); // TODO: think about handling exception here
     }
 
-    public void deleteUserByUsername (String username) {
-        List<Customer> customerUsername = customerRepository.findByUserName(username);
-        if (!customerUsername.isEmpty()) {
-            noteRepository.deleteAllById(List.of(customerUsername.getFirst().getId()));
-            customerRepository.deleteById(customerUsername.getFirst().getId());
+    public void deleteUserByUsername (String username, Customer requester) throws IllegalAccessException {
+        if (requester.getRoles().stream()
+                .map(Role::getName)
+                .anyMatch(roleName -> "ROLE_ADMIN".equals(roleName))) {
+            List<Customer> customerUsername = customerRepository.findByUsername(username);
+            if (!customerUsername.isEmpty()) {
+                noteRepository.deleteAllById(List.of(customerUsername.getFirst().getId()));
+                customerRepository.deleteById(customerUsername.getFirst().getId());
+            }
+            else {
+                throw new UsernameNotFoundException("User not found");
+            }
         }
         else {
-            throw new UsernameNotFoundException("User not found");
+            throw new IllegalAccessException("You do not have access to delete a user");
         }
+    }
+
+    public void deleteUserById (Long id, Customer requester) throws IllegalAccessException, UserPrincipalNotFoundException {
+        if (requester.getRoles().stream()
+                .map(Role::getName)
+                .anyMatch(roleName -> "ROLE_ADMIN".equals(roleName))) {
+            Optional<Customer> customer = customerRepository.findById(id);
+            if (customer.isPresent()) {
+                noteRepository.deleteAllById(List.of(customer.get().getId()));
+                customerRepository.deleteById(customer.get().getId());
+            }
+            else {
+                throw new UserPrincipalNotFoundException("User not found");
+            }
+        }
+        else {
+            throw new IllegalAccessException("You do not have access to delete a user");
+        }
+    }
+
+    public Iterable<Customer> getAllCustomers (Customer requester) throws IllegalAccessException {
+        if (requester.getRoles().stream()
+                .map(Role::getName)
+                .anyMatch(roleName -> "ROLE_ADMIN".equals(roleName))) {
+            log.info("{} as ADMIN just requested all customers", requester.getUsername());
+            return customerRepository.findAll();
+        } else {
+            log.error("The caller: {} does not have admin rights to list all users", requester.getUsername());
+            throw new IllegalAccessException("You do not have access to list all users");
+        }
+    }
+
+    public boolean isAdmin (Customer customer) {
+        if (customer.getRoles().stream()
+                .map(Role::getName)
+                .anyMatch(roleName -> "ROLE_ADMIN".equals(roleName)))
+            return true;
+        else return false;
     }
 }
